+++ NEWS.md
@@ -1,5 +1,6 @@
 # medfateland 2.4.6
 - Numerical species codes allowed as input for fordyn_land()
+- Added filterMissingSpecies in parse_forestable()
 
 # medfateland 2.4.5
 - New helper function parse_forestable()